move 3 functionality tests into a unit test
